membership_token/pyproject.toml
[tool.poetry]
name = "membership"
version = "0.1.0"
description = ""
authors = ["Gomnonix <lee20481997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
